requirements.txt
opencv-python==4.6.0.66
Pillow==9.2.0
face_recognition==1.3.0

face-recognition~=1.3.0
numpy~=1.18.5
flask~=3.0.3
setuptools~=75.3.0
torch~=2.4.1+cu118
torchvision~=0.19.1+cu118
tqdm~=4.31.1
tensorboardX~=2.0
easydict~=1.9